+++ src/test/java/net/sf/jabref/logic/integrity/IntegrityCheckTest.java
@@ -112,6 +112,12 @@ public class IntegrityCheckTest {
     }
 
     @Test
+    public void testBibtexkeyChecks() {
+        assertCorrect(createContext("bibtexkey", "Knuth2014"));
+        assertCorrect(createContext("bibtexkey", ""));
+    }
+
+    @Test
     public void testBracketChecks() {
         assertCorrect(createContext("title", "x"));
         assertCorrect(createContext("title", "{x}"));